Initial implementation of an actual block back driver in user space. Testing has been occuring on Xen-based Beam with an Ubuntu Linux HVM serving as the storage domain. The guest is a stripped down XSVM that only has enough in it to perform a switch root essentially.
Able to mount a filesystem via this block back if A LOT of caching is disabled
1) mount the image on a loop device:
  $ losetup -fP -b 512 -o 0 --direct-io=on /storage/ndvm.raw
2) within the guest:
  $ echo 0 > /sys/block/xvda/queue/read_ahead_kb
  $ echo none > /sys/block/xvda/queue/write_cache
3) pray to the ext4 gods
4) Within the guest:
  $ mount -o sync,inode_readahead_blks=0,nombcache /dev/xvda /mnt

I implemented a lot more of the protocol than I intended to due to some
unexpected behavior by the frontend when the most basic version of the
protocol was used. This should help improve throughput for reads and
writes.

Signed-off-by: Brendan Kerrigan <kerriganb@ainfosec.com>