Add code for running as a Windows service
When passed --windows-svc, the backend will be run in a Windows
service. This service (as of now) only accepts the PRESHUTDOWN
control (i.e. it cant be stopped). This allows it to be used as
a dependency for VMs that need persistent storage throughout their
lifetime, without risk of their disk backend disappearing.