Use splice for LRU maintenance
Use the splice member to move the accessed gref to the MRU position
within the LRU list (i.e. the front) instead of push_front+erase.
This eliminates a call to malloc and to free each time a gref
is accessed, since splice simply adjusts the pointers of the existing
list entries.